+++ HISTORY.md
@@ -5,9 +5,13 @@
 * **Notable Fixes:**
     * fixed stretching SVG costumes with fixed aspect ratios in Firefox
     * only report neighbors that are visible
+* **Translation Updates:**
+    * Italian, thanks, Stefano!
+
 
 ### 2020-11-20
 * threads: only report neighbors that are visible, thanks Frederic, for reporting this bug!
+* Italian translation update, thanks, Stefano!
 
 ### 2020-11-19
 * new dev version